pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=4.3.0",
    "fastapi>=0.115.14",
    "passlib>=1.7.4",
    "pydantic>=2.11.7",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "uvicorn>=0.35.0",
]
